+++ spec/unit/imap/backup/serializer_spec.rb
@@ -1,4 +1,21 @@
 shared_examples "a method that checks for invalid serialization" do
+  require "imap/backup/serializer/version2_migrator"
+
+  context "with version 2 metadata files" do
+    let(:version2_migrator) do
+      instance_double(Imap::Backup::Serializer::Version2Migrator, run: false)
+    end
+
+    before do
+      allow(Imap::Backup::Serializer::Version2Migrator).to receive(:new) { version2_migrator }
+      action.call
+    end
+
+    it "migrates to version 3" do
+      expect(version2_migrator).to have_received(:run)
+    end
+  end
+
   context "when either file is invalid" do
     let(:imap_valid) { true }
     let(:mbox_valid) { true }